scripts renaming
* processa_meta_whats_logs => process_html_logs_extractions_to_text
