+++ scripts/process_html_logs_extractions_to_text.py
